Refactored logic to easily add new indices.
