+++ index-rebalances/assetInfo.ts
@@ -12,7 +12,7 @@ export interface Assets {
   [symbol: string]: AssetInfo;
 }
 
-export const assets: Assets = {
+export const ASSETS: Assets = {
   YFI: {
     id: "yfi",
     address: "0x0bc529c00C6401aEF6D220BE8C6Ea1667F6Ad93e",
