+++ index-rebalances/dependencies.ts
@@ -1,6 +1,4 @@
 export default {
-  DPI: "0x73BAA8A41ddA1Cbcae8C7709A2A7B171182A1D46",
-  MVI: "0x72e364F2ABdC788b7E918bc238B21f109Cd634D7",
   GENERAL_INDEX_MODULE: "0x776c27990Ce03D094Ff2241Ff8229cF98b91364D",
   DPI_SINGLE_INDEX_MODULE: "0x25100726b25a6ddb8f8e68988272e1883733966e",
 }
