+++ index-rebalances/indices/dpi/strategyInfo.ts
@@ -1,12 +1,12 @@
 import { BigNumber } from "@ethersproject/bignumber";
-import { ZERO } from "../../utils/constants";
-import { ether } from "../../utils/common/index";
-import { assets } from "../assetInfo"
-import { exchanges, StrategyInfo } from "../types";
+import { ZERO } from "../../../utils/constants";
+import { ether } from "../../../utils/common/index";
+import { ASSETS } from "../../assetInfo"
+import { exchanges, StrategyInfo } from "../../types";
 
 export const strategyInfo: StrategyInfo = {
   YFI: {
-    address: assets.YFI.address,
+    address: ASSETS.YFI.address,
     supply: BigNumber.from(36096),
     // maxTradeSize: ether(13.2),
     maxTradeSize: ether(1),
@@ -15,7 +15,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   COMP: {
-    address: assets.COMP.address,
+    address: ASSETS.COMP.address,
     supply: BigNumber.from(5048104),
     // maxTradeSize: ether(192),
     maxTradeSize: ether(0.03),
@@ -24,7 +24,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   SNX: {
-    address: assets.SNX.address,
+    address: ASSETS.SNX.address,
     supply: BigNumber.from(151684080),
     // maxTradeSize: ether(10250),
     maxTradeSize: ether(1),
@@ -33,7 +33,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   MKR: {
-    address: assets.MKR.address,
+    address: ASSETS.MKR.address,
     supply: BigNumber.from(901683),
     // maxTradeSize: ether(40),
     maxTradeSize: ether(1),
@@ -42,7 +42,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   REN: {
-    address: assets.REN.address,
+    address: ASSETS.REN.address,
     supply: BigNumber.from(881153018),
     // maxTradeSize: ether(55000),
     maxTradeSize: ether(1),
@@ -51,7 +51,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   KNC: {
-    address: assets.KNC.address,
+    address: ASSETS.KNC.address,
     supply: BigNumber.from(204614446),
     // maxTradeSize: ether(2100),
     maxTradeSize: ether(12),
@@ -60,7 +60,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   LRC: {
-    address: assets.LRC.address,
+    address: ASSETS.LRC.address,
     supply: BigNumber.from(1245991468),
     // maxTradeSize: ether(14000),
     maxTradeSize: ether(20),
@@ -69,7 +69,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   BAL: {
-    address: assets.BAL.address,
+    address: ASSETS.BAL.address,
     supply: BigNumber.from(10799858),
     // maxTradeSize: ether(9900),
     maxTradeSize: ether(1),
@@ -78,7 +78,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   UNI: {
-    address: assets.UNI.address,
+    address: ASSETS.UNI.address,
     supply: BigNumber.from(519857388),
     // maxTradeSize: ether(17000),
     maxTradeSize: ether(1),
@@ -87,7 +87,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   AAVE: {
-    address: assets.AAVE.address,
+    address: ASSETS.AAVE.address,
     supply: BigNumber.from(12492713),
     // maxTradeSize: ether(1200),
     maxTradeSize: ether(1),
@@ -96,7 +96,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   MTA: {
-    address: assets.MTA.address,
+    address: ASSETS.MTA.address,
     supply: BigNumber.from(24894863),
     // maxTradeSize: ether(5000),
     maxTradeSize: ether(10),
@@ -105,7 +105,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   SUSHI: {
-    address: assets.SUSHI.address,
+    address: ASSETS.SUSHI.address,
     supply: BigNumber.from(148978846),
     // maxTradeSize: ether(72000),
     maxTradeSize: ether(2),
@@ -114,7 +114,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   CREAM: {
-    address: assets.CREAM.address,
+    address: ASSETS.CREAM.address,
     supply: BigNumber.from(685568),
     // maxTradeSize: ether(135),
     maxTradeSize: ether(1),
@@ -123,7 +123,7 @@ export const strategyInfo: StrategyInfo = {
     currentUnit: ZERO,
   },
   FARM: {
-    address: assets.FARM.address,
+    address: ASSETS.FARM.address,
     supply: BigNumber.from(528505),
     // maxTradeSize: ether(125),
     maxTradeSize: ether(1),
