+++ package.json
@@ -16,7 +16,7 @@
   ],
   "scripts": {
     "build": "yarn compile && yarn patch-hardhat-typechain && yarn typechain && yarn fix-typechain && yarn transpile-dist",
-    "calculate-dpi-units": "npx hardhat --network production calculate-new-dpi-position",
+    "calculate-index-units": "npx hardhat --network production calculate-new-index-position",
     "validate-dpi-rebalance": "npx hardhat --network production validate-dpi-params",
     "chain": "npx hardhat node --no-deploy",
     "clean": "rm -f coverage.json; rm -rf .coverage_cache; rm -rf .coverage_contracts; rm -rf cache; rm -rf coverage; rm -rf typechain; rm -rf artifacts",
