+++ tasks/index.ts
@@ -1,3 +1,3 @@
 export * from "./subtasks";
-export * from "./calculate-new-dpi-position";
+export * from "./calculate-new-index-position";
 export * from "./validate-dpi-params";
