+++ utils/contracts/setV2.ts
@@ -8,6 +8,7 @@ export { ContractCallerMock } from "../../typechain/ContractCallerMock";
 export { ComptrollerMock } from "../../typechain/ComptrollerMock";
 export { CompoundLeverageModule } from "../../typechain/CompoundLeverageModule";
 export { DebtIssuanceModule } from "../../typechain/DebtIssuanceModule";
+export { GeneralIndexModule } from "../../typechain/GeneralIndexModule";
 export { IntegrationRegistry } from "../../typechain/IntegrationRegistry";
 export { SingleIndexModule } from "../../typechain/SingleIndexModule";
 export { StreamingFeeModule } from "../../typechain/StreamingFeeModule";
